+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added an inverted look camera option (#700)
 - added a camera speed option for the manual camera (#815)
 - added an option to fix original texture issues (#826)
+- added menu specific controls (#814)
 - fixed sounds stopping instead of pausing if game sounds in inventory are disabled (#717)
 - fixed ceiling heights at times being miscalculated, resulting in camera issues and Lara being able to jump into the ceiling (#323)
 - fixed Lara not being able to jump off trapdoors or crumbling floors if the sidestep descent fix is enabled (#830)